Fix concurrency issue in Http4sCacheSessionStore.
Using a non-concurrent mutable hash map here can lead to infinite loops
when multiple logins occur simultaneously.

Ironically, the logging may be hiding this issue a bit, as when there is
a logger, the writing of it can block and act a bit like a synchronise.

We've hit this in prod, and it was taking down the server.